requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
crispy-bootstrap5==2023.10
djangify==1.0.0
Django==5.0.2
django-crispy-forms==2.1
djangorestframework==3.14.0
image==1.5.33
pillow==10.2.0
psycopg2==2.9.9
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
